Implement GPOS1.2 and split GPOS1 out to own file
